+++ R/model.R
@@ -887,19 +887,20 @@ anc_testing_artcov_mf <- function(year, anc_testing, naomi_mf) {
 #' @rdname anc_testing_prev_mf
 #'
 #' @param calendar_quarter Calendar quarter
+#'
+#' @details
+#'
+#' Number on ART at desired quarter are linearly interpolated within the dataset.
+#' If the desired quarter is before the earliest data, the first value may
+#' be carried back by up to one year (four quarters). Data are never carried forward
+#' 
 #' @export
 artnum_mf <- function(calendar_quarter, art_number, naomi_mf) {
 
   stopifnot(length(calendar_quarter) <= 1)
   stopifnot(is(naomi_mf, "naomi_mf"))
 
-  if(!is.null(calendar_quarter)) {
-    year <- year_labels(calendar_quarter_to_quarter_id(calendar_quarter))
-  } else {
-    year <- NULL
-  }
-
-  if(is.null(year) || is.null(art_number)) {
+  if(is.null(calendar_quarter) || is.null(art_number)) {
     ## No number on ART data or no year specified
 
     artnum_dat <- data.frame(
@@ -912,16 +913,31 @@ artnum_mf <- function(calendar_quarter, art_number, naomi_mf) {
     )
   } else {
 
-    art_number$year <- year_labels(calendar_quarter_to_quarter_id(art_number$calendar_quarter))
+    out_quarter_id <- calendar_quarter_to_quarter_id(calendar_quarter)
     
-    if(!year %in% art_number$year)
-      stop(paste0("No ART data found for year ", year, ".\n",
-                  "Set year = NULL if you intend to include no ART data."))
+    dat <- art_number %>%
+      dplyr::inner_join(
+               dplyr::select(naomi_mf$mf_areas, area_id, spectrum_region_code),
+               by = "area_id"
+             ) %>%
+      dplyr::mutate(
+               quarter_id = calendar_quarter_to_quarter_id(calendar_quarter)
+             )
     
-    ## !!! Note: should add some subsetting for sex and age group.
-    artnum_dat <- art_number %>%
-      dplyr::filter(year == !!year,
-                    area_id %in% naomi_mf$mf_areas$area_id) %>%
+    dat <- dat %>%
+      dplyr::group_by(area_id, sex, age_group, spectrum_region_code) %>%
+      dplyr::summarise(min_data_quarter = min(quarter_id),
+                       max_data_quarter = max(quarter_id),
+                       current_art = approx(quarter_id, current_art, out_quarter_id, rule =2)$y) %>%
+      dplyr::ungroup() %>%
+      dplyr::filter(out_quarter_id > min_data_quarter - 4L,
+                    out_quarter_id <= max_data_quarter)
+
+    if(nrow(dat) == 0)
+      stop(paste0("No ART data found for quarter ", calendar_quarter, ".\n", 
+            "Set calendar_quarter = NULL if you intend to include no ART data."))
+    
+    artnum_dat <- dat %>%
       dplyr::transmute(
                area_id,
                sex,
@@ -930,6 +946,6 @@ artnum_mf <- function(calendar_quarter, art_number, naomi_mf) {
                current_art
              )
   }
-
+  
   artnum_dat
 }