+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add spatial interaction to ART coverage model for <15 / 15+ to allow different paediatric vs. adult ART coverage and different change in paediatric coverage. This reduces paediatric ART data distorting adult model results.
 * ART number dataset accepts either year or calendar_quarter column.
+* Linearly interpolate number on ART for model fitting. If desired quarter is before earliest data, the earliest ART number may be carried backward by up to four quarters. Number on ART are never carried forwards.
 
 # naomi 0.0.49